refactor: restructure agui modules to match slack pattern
- Merge completion.py into events.py (all event creation in one place)
- Rename messages.py → helpers.py (add validation + formatting)
- Rename streaming.py → handler.py (matches slack's event_handler.py)
- Simplify state.py to EventBuffer only
- Unify run_agent/run_team into run_entity in router.py
- Consolidate duplicate tests

Final structure (4 core modules):
- state.py: EventBuffer per-stream state
- events.py: chunk/completion/delta event creation
- handler.py: sync/async stream orchestration
- helpers.py: extract, validate, format utilities
