+++ libs/agno/agno/os/interfaces/agui/helpers.py
@@ -1,11 +1,15 @@
-"""Text extraction helpers for AG-UI user messages and Agno response chunks."""
+"""Utility functions for AG-UI: message extraction, state validation, formatting."""
 
-from typing import List
+from dataclasses import asdict, is_dataclass
+from typing import Any, Dict, List, Optional
 
 from ag_ui.core.types import Message as AGUIMessage
+from pydantic import BaseModel
 
+from agno.reasoning.step import ReasoningStep
 from agno.run.agent import RunContentEvent
 from agno.run.team import RunContentEvent as TeamRunContentEvent
+from agno.utils.log import log_warning
 from agno.utils.message import get_text_from_message
 
 
@@ -18,10 +22,8 @@ def extract_agui_user_input(messages: List[AGUIMessage]) -> str:
     """
     for msg in reversed(messages):
         if msg.role == "user" and msg.content is not None:
-            # UserMessage.content is Union[str, List[InputContent]]
             if isinstance(msg.content, str):
                 return msg.content
-            # Multimodal: extract text parts
             if isinstance(msg.content, list):
                 text_parts = []
                 for part in msg.content:
@@ -34,14 +36,10 @@ def extract_agui_user_input(messages: List[AGUIMessage]) -> str:
 
 def extract_response_chunk_content(response: RunContentEvent) -> str:
     """Given a response stream chunk, find and extract the content."""
-
     if hasattr(response, "messages") and response.messages:  # type: ignore
         for msg in reversed(response.messages):  # type: ignore
             if hasattr(msg, "role") and msg.role == "assistant" and hasattr(msg, "content") and msg.content:
-                # Handle structured outputs from messages
                 return get_text_from_message(msg.content)
-
-    # Handle structured outputs
     return get_text_from_message(response.content) if response.content is not None else ""
 
 
@@ -51,8 +49,6 @@ def extract_team_response_chunk_content(response: TeamRunContentEvent) -> str:
     Recursively folds in member responses so the AG-UI client sees a single
     consolidated text instead of N separate member text deltas.
     """
-
-    # Handle Team members' responses
     members_content = []
     if hasattr(response, "member_responses") and response.member_responses:  # type: ignore
         for member_resp in response.member_responses:  # type: ignore
@@ -65,8 +61,63 @@ def extract_team_response_chunk_content(response: TeamRunContentEvent) -> str:
                 if member_content:
                     members_content.append(f"Team member: {member_content}")
     members_response = "\n".join(members_content) if members_content else ""
-
-    # Handle structured outputs
     main_content = get_text_from_message(response.content) if response.content is not None else ""
-
     return main_content + members_response
+
+
+def validate_agui_state(state: Any, thread_id: str) -> Optional[Dict[str, Any]]:
+    """Validate the given AGUI state is of the expected type (dict)."""
+    if state is None:
+        return None
+
+    if isinstance(state, dict):
+        return state
+
+    if isinstance(state, BaseModel):
+        try:
+            return state.model_dump()
+        except Exception:
+            pass
+
+    if is_dataclass(state):
+        try:
+            return asdict(state)  # type: ignore
+        except Exception:
+            pass
+
+    if hasattr(state, "to_dict") and callable(getattr(state, "to_dict")):
+        try:
+            result = state.to_dict()  # type: ignore
+            if isinstance(result, dict):
+                return result
+        except Exception:
+            pass
+
+    log_warning(f"AGUI state must be a dict, got {type(state).__name__}. State will be ignored. Thread: {thread_id}")
+    return None
+
+
+def format_reasoning_step_delta(step: Optional[ReasoningStep], step_number: int = 0) -> str:
+    """Format a single ReasoningStep as a text delta for REASONING_MESSAGE_CONTENT.
+
+    ReasoningStepEvent.content holds a ReasoningStep object (title, reasoning,
+    action, result, confidence). We format just this one step — NOT the
+    accumulated reasoning_content field, which duplicates prior steps.
+    """
+    if step is None:
+        return ""
+    parts: List[str] = []
+    title = step.title or "Thinking"
+    if step_number > 0:
+        parts.append(f"## Step {step_number}: {title}")
+    else:
+        parts.append(f"## {title}")
+    if step.reasoning:
+        parts.append(step.reasoning)
+    if step.action:
+        parts.append(f"Action: {step.action}")
+    if step.result:
+        parts.append(f"Result: {step.result}")
+    if step.confidence is not None:
+        parts.append(f"Confidence: {step.confidence}")
+    return "\n".join(parts) + "\n\n" if parts else ""
