+++ libs/agno/agno/os/interfaces/agui/router.py
@@ -23,41 +23,35 @@ from fastapi import APIRouter
 from fastapi.responses import StreamingResponse
 
 from agno.agent import Agent, RemoteAgent
+from agno.os.interfaces.agui.handler import async_stream_agno_response_as_agui_events
+from agno.os.interfaces.agui.helpers import extract_agui_user_input, validate_agui_state
 from agno.os.interfaces.agui.media import extract_agui_media
-from agno.os.interfaces.agui.messages import extract_agui_user_input
-from agno.os.interfaces.agui.state import validate_agui_state
-from agno.os.interfaces.agui.streaming import async_stream_agno_response_as_agui_events
 from agno.team.remote import RemoteTeam
 from agno.team.team import Team
 
 
-async def run_agent(agent: Union[Agent, RemoteAgent], run_input: RunAgentInput) -> AsyncIterator[BaseEvent]:
-    """Run the contextual Agent, mapping AG-UI input messages to Agno format, and streaming the response in AG-UI format."""
+async def run_entity(
+    entity: Union[Agent, RemoteAgent, Team, RemoteTeam],
+    run_input: RunAgentInput,
+) -> AsyncIterator[BaseEvent]:
+    """Shared handler for running an Agent or Team with AG-UI input/output mapping."""
     run_id = run_input.run_id or str(uuid.uuid4())
 
     try:
         # AG-UI frontends send full conversation history every request.
-        # Extract only the last user message — agent manages history via session DB.
+        # Extract only the last user message — entity manages history via session DB.
         user_input = extract_agui_user_input(run_input.messages or [])
         images, audio, videos, files = extract_agui_media(run_input.messages or [])
 
         yield RunStartedEvent(type=EventType.RUN_STARTED, thread_id=run_input.thread_id, run_id=run_id)
 
-        # Look for user_id in run_input.forwarded_props
-        user_id = None
-        if run_input.forwarded_props and isinstance(run_input.forwarded_props, dict):
-            user_id = run_input.forwarded_props.get("user_id")
-
-        # Validating the session state is of the expected type (dict)
+        user_id = run_input.forwarded_props.get("user_id") if run_input.forwarded_props else None
         session_state = validate_agui_state(run_input.state, run_input.thread_id)
 
-        # Emit initial state snapshot if state is provided
         if session_state is not None:
-            # Deep-copy so the emitted event doesn't alias the live agent state (consistent with final snapshot).
             yield StateSnapshotEvent(type=EventType.STATE_SNAPSHOT, snapshot=copy.deepcopy(session_state))
 
-        # Request streaming response from agent
-        response_stream = agent.arun(  # type: ignore
+        response_stream = entity.arun(  # type: ignore
             input=user_input,
             session_id=run_input.thread_id,
             stream=True,
@@ -71,7 +65,6 @@ async def run_agent(agent: Union[Agent, RemoteAgent], run_input: RunAgentInput) 
             run_id=run_id,
         )
 
-        # Stream the response content in AG-UI format
         async for event in async_stream_agno_response_as_agui_events(
             response_stream=response_stream,  # type: ignore
             thread_id=run_input.thread_id,
@@ -80,58 +73,8 @@ async def run_agent(agent: Union[Agent, RemoteAgent], run_input: RunAgentInput) 
         ):
             yield event
 
-    # Emit a RunErrorEvent if any error occurs
     except Exception as e:
-        log_error(f"Error running agent: {str(e)}")
-        yield RunErrorEvent(type=EventType.RUN_ERROR, message=str(e))
-
-
-async def run_team(team: Union[Team, RemoteTeam], input: RunAgentInput) -> AsyncIterator[BaseEvent]:
-    """Run the contextual Team, mapping AG-UI input messages to Agno format, and streaming the response in AG-UI format."""
-    run_id = input.run_id or str(uuid.uuid4())
-    try:
-        # AG-UI frontends send full conversation history every request.
-        # Extract only the last user message — team manages history via session DB.
-        user_input = extract_agui_user_input(input.messages or [])
-        images, audio, videos, files = extract_agui_media(input.messages or [])
-        yield RunStartedEvent(type=EventType.RUN_STARTED, thread_id=input.thread_id, run_id=run_id)
-
-        # Look for user_id in input.forwarded_props
-        user_id = None
-        if input.forwarded_props and isinstance(input.forwarded_props, dict):
-            user_id = input.forwarded_props.get("user_id")
-
-        # Validating the session state is of the expected type (dict)
-        session_state = validate_agui_state(input.state, input.thread_id)
-
-        # Emit initial state snapshot if state is provided
-        if session_state is not None:
-            # Deep-copy so the emitted event doesn't alias the live agent state (consistent with final snapshot).
-            yield StateSnapshotEvent(type=EventType.STATE_SNAPSHOT, snapshot=copy.deepcopy(session_state))
-
-        # Request streaming response from team
-        response_stream = team.arun(  # type: ignore
-            input=user_input,
-            session_id=input.thread_id,
-            stream=True,
-            stream_events=True,
-            user_id=user_id,
-            images=images or None,
-            audio=audio or None,
-            videos=videos or None,
-            files=files or None,
-            session_state=session_state,
-            run_id=run_id,
-        )
-
-        # Stream the response content in AG-UI format
-        async for event in async_stream_agno_response_as_agui_events(
-            response_stream=response_stream, thread_id=input.thread_id, run_id=run_id, run_state=session_state
-        ):
-            yield event
-
-    except Exception as e:
-        log_error(f"Error running team: {str(e)}")
+        log_error(f"Error running entity: {str(e)}")
         yield RunErrorEvent(type=EventType.RUN_ERROR, message=str(e))
 
 
@@ -141,22 +84,14 @@ def attach_routes(
     if agent is None and team is None:
         raise ValueError("Either agent or team must be provided.")
 
+    entity = agent or team
     encoder = EventEncoder()
 
-    @router.post(
-        "/agui",
-        name="run_agent",
-    )
+    @router.post("/agui", name="run_agent")
     async def run_agent_agui(run_input: RunAgentInput):
         async def event_generator():
-            if agent:
-                async for event in run_agent(agent, run_input):
-                    encoded_event = encoder.encode(event)
-                    yield encoded_event
-            elif team:
-                async for event in run_team(team, run_input):
-                    encoded_event = encoder.encode(event)
-                    yield encoded_event
+            async for event in run_entity(entity, run_input):  # type: ignore
+                yield encoder.encode(event)
 
         return StreamingResponse(
             event_generator(),
