+++ libs/agno/agno/os/interfaces/agui/state.py
@@ -1,62 +1,25 @@
-"""Per-stream state buffer, validation, and JSON-Patch delta computation for the AG-UI interface."""
+"""Per-stream state buffer for AG-UI event ordering."""
 
 import copy
 import uuid
-from dataclasses import asdict, dataclass, field, is_dataclass
+from dataclasses import dataclass, field
 from typing import Any, Dict, List, Optional, Set, Tuple
 
-from ag_ui.core import BaseEvent, EventType, StateDeltaEvent
-from pydantic import BaseModel
-
 from agno.utils.log import log_warning
-
-
-def validate_agui_state(state: Any, thread_id: str) -> Optional[Dict[str, Any]]:
-    """Validate the given AGUI state is of the expected type (dict)."""
-    if state is None:
-        return None
-
-    if isinstance(state, dict):
-        return state
-
-    if isinstance(state, BaseModel):
-        try:
-            return state.model_dump()
-        except Exception:
-            pass
-
-    if is_dataclass(state):
-        try:
-            return asdict(state)  # type: ignore
-        except Exception:
-            pass
-
-    if hasattr(state, "to_dict") and callable(getattr(state, "to_dict")):
-        try:
-            result = state.to_dict()  # type: ignore
-            if isinstance(result, dict):
-                return result
-        except Exception:
-            pass
-
-    log_warning(f"AGUI state must be a dict, got {type(state).__name__}. State will be ignored. Thread: {thread_id}")
-    return None
 
 
 @dataclass
 class EventBuffer:
     """Buffer to manage event ordering constraints, relevant when mapping Agno responses to AG-UI events."""
 
-    active_tool_call_ids: Set[str]  # All currently active tool calls
-    ended_tool_call_ids: Set[str]  # All tool calls that have ended
-    current_text_message_id: str = ""  # ID of the current text message context (for tool call parenting)
-    next_text_message_id: str = ""  # Pre-generated ID for the next text message
-    pending_tool_calls_parent_id: str = ""  # Parent message ID for pending tool calls
-    reasoning_message_id: Optional[str] = None  # Active reasoning session ID, set by reasoning_started
-    reasoning_step_count: int = 0  # Step counter for ReasoningTools (reset each session)
-    _last_snapshot: Optional[Dict[str, Any]] = field(
-        default=None, repr=False
-    )  # Snapshot of last state for delta computation
+    active_tool_call_ids: Set[str]
+    ended_tool_call_ids: Set[str]
+    current_text_message_id: str = ""
+    next_text_message_id: str = ""
+    pending_tool_calls_parent_id: str = ""
+    reasoning_message_id: Optional[str] = None
+    reasoning_step_count: int = 0
+    _last_snapshot: Optional[Dict[str, Any]] = field(default=None, repr=False)
 
     def __init__(self):
         self.active_tool_call_ids = set()
@@ -69,49 +32,38 @@ class EventBuffer:
         self._last_snapshot = None
 
     def start_tool_call(self, tool_call_id: str) -> None:
-        """Start a new tool call."""
         self.active_tool_call_ids.add(tool_call_id)
 
     def end_tool_call(self, tool_call_id: str) -> None:
-        """End a tool call."""
         self.active_tool_call_ids.discard(tool_call_id)
         self.ended_tool_call_ids.add(tool_call_id)
 
     def start_text_message(self) -> str:
-        """Start a new text message and return its ID."""
         self.current_text_message_id = self.next_text_message_id
         self.next_text_message_id = str(uuid.uuid4())
         return self.current_text_message_id
 
     def get_parent_message_id_for_tool_call(self) -> str:
-        """Get the message ID to use as parent for tool calls."""
         if self.pending_tool_calls_parent_id:
             return self.pending_tool_calls_parent_id
         return self.current_text_message_id
 
     def set_pending_tool_calls_parent_id(self, parent_id: str) -> None:
-        """Set the parent message ID for upcoming tool calls."""
         self.pending_tool_calls_parent_id = parent_id
 
     def clear_pending_tool_calls_parent_id(self) -> None:
-        """Clear the pending parent ID when a new text message starts."""
         self.pending_tool_calls_parent_id = ""
 
     def start_reasoning(self) -> str:
-        """Start a new reasoning session and return its message ID."""
         self.reasoning_message_id = str(uuid.uuid4())
         self.reasoning_step_count = 0
         return self.reasoning_message_id
 
     def set_state_snapshot(self, state: Dict[str, Any]) -> None:
-        """Store deep copy of current state for delta computation."""
         self._last_snapshot = copy.deepcopy(state)
 
     def compute_state_delta(self, current_state: Dict[str, Any]) -> Optional[List[Dict[str, Any]]]:
-        """Compute JSON Patch delta between last snapshot and current state.
-
-        Returns a list of JSON Patch ops (RFC 6902) or None if unchanged/error.
-        """
+        """Compute JSON Patch delta between last snapshot and current state."""
         if self._last_snapshot is None:
             return None
         try:
@@ -127,35 +79,15 @@ class EventBuffer:
             return None
 
     def next_reasoning_step(self) -> int:
-        """Increment and return the current reasoning step number."""
         self.reasoning_step_count += 1
         return self.reasoning_step_count
 
     def ensure_reasoning_started(self) -> Tuple[str, bool]:
-        """Return the active reasoning session ID, starting one if needed.
-
-        Returns (reasoning_id, is_new) where is_new is True if a new session was created.
-        """
+        """Return the active reasoning session ID, starting one if needed."""
         if self.reasoning_message_id is not None:
             return self.reasoning_message_id, False
         return self.start_reasoning(), True
 
     def end_reasoning(self) -> None:
-        """End the active reasoning session."""
         self.reasoning_message_id = None
         self.reasoning_step_count = 0
-
-
-def create_state_delta_events(
-    run_state: Optional[Dict[str, Any]],
-    event_buffer: EventBuffer,
-) -> List[BaseEvent]:
-    """Compute state delta and return StateDeltaEvent if state changed."""
-    if run_state is None:
-        return []
-    ops = event_buffer.compute_state_delta(run_state)
-    if ops is None:
-        return []
-    # Update the snapshot to current state for next delta computation
-    event_buffer.set_state_snapshot(run_state)
-    return [StateDeltaEvent(type=EventType.STATE_DELTA, delta=ops)]
